E-17: extract SignatureEncoder from ReflectionMetadataEmitter (#1361)
Peels the signature / type-encoding band — the core of the emitter's
metadata output — off ReflectionMetadataEmitter into a new internal sealed
SignatureEncoder (back-reference idiom, ctor(outer), convenience fields for
emitCtx/cache/remaps). RME drops from 11,122 to 10,104 lines; SignatureEncoder
is 1,152.

Band moved (verbatim except mechanical root-member → outer/emitCtx/cache/remaps
rewrites): EncodeTypeSymbol, EncodeClrType, EncodeReturnSymbol (both overloads),
EncodeReturnClr, EncodeLocalVariableType, EncodeFunctionTypeSymbol,
GetFunctionPointerCallSiteSignature, MapToSignatureCallingConvention, and the
host→reference-context CLR type resolution the reflection encoding path depends
on: ResolveTargetDelegateClrType, ResolveDelegateClrType, ResolveDelegateArgClrType,
ResolveAsyncDelegateClrType, MapToReferenceClrType.

Boundary helpers left on root (call the moved encoders through the forwarders,
and are themselves consumed by the not-yet-extracted token-resolution band that
moves in E-18/E-19): EncodeTypeSymbolIntoSignature, EncodeFunctionReturnSymbol,
EncodeOpenMethodSignature, EncodeAsyncReturnType, the delegate MemberRef/TypeSpec
producers (GetFunctionDelegateTypeSpec/CtorRef/InvokeRef, GetConstructedDelegateCtorRef,
TryGetSymbolicAsyncDelegateCtorRef) and their caches, the async-result predicates
IsAsyncUserDefinedResultType / TryCreateSymbolicAsyncTaskType, and the widely-shared
value-type / symbolic predicates (IsValueTypeSymbol, ArgIsSymbolicUserDefined,
FunctionTypeNeedsSymbolicDelegate, IsValueTypeNullable).

Forwarders added on root for E-21 cleanup (keep the ~90 in-RME call sites and the
external outer.<Encode…> call sites in MethodBodyEmitter.* compiling untouched):
EncodeTypeSymbol, EncodeReturnSymbol, EncodeReturnClr, EncodeClrType,
EncodeLocalVariableType, EncodeFunctionTypeSymbol, GetFunctionPointerCallSiteSignature,
ResolveTargetDelegateClrType, ResolveDelegateClrType, ResolveAsyncDelegateClrType,
MapToReferenceClrType.

Collaborator-injection sites in EmitCore rebound directly to signatures.*:
EncodeTypeSymbol (methodBodyPlanner, dataStructSynth, memberDefEmitter,
typeDefEmitter, stateMachines, TypeDefEmitter.EncodeParameterSignature call),
EncodeReturnSymbol (typeDefEmitter), EncodeClrType (stateMachines).

Temporary outer-coupling to user-token resolvers (resolved in E-18/E-19):
the band calls outer.GetTypeReference, outer.ResolveUserStructTypeSpecArguments
(widened private → internal), ReflectionMetadataEmitter.ResolveDelegateTypeArguments
(widened private → internal, static), and outer.stateMachines.AsyncStateMachinePlans.
The active-remap reads route through remaps.ActiveIteratorStateMachineRemap /
ActiveLambdaMethodTypeParamRemap and are read (never snapshotted), preserving the
object-identity cache-key semantics.

Verification: dotnet build GSharp.sln -c Release -graph → 0 warnings, 0 errors.
PE byte-diff over samples/*.gs: 119/119 emitted PEs byte-for-byte identical to the
pre-change baseline (3 GsharpExtensions* baseline failures skipped).
